renamed component to pass lint line limit.
