+++ public-ui/src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import {NavigationEnd, Router, RouterModule, Routes} from '@angular/router';
 
 import {DbHomeComponent} from './views/db-home/db-home.component';
 import {EhrViewComponent} from './views/ehr-view/ehr-view.component';
-import {PhysicalMeasurementsComponent} from './views/physical-measurements/physical-measurements.component';
+import {PhysicalMeasurementsComponent} from './views/pm/pm.component';
 import {QuickSearchComponent} from './views/quick-search/quick-search.component';
 import {SurveyViewComponent} from './views/survey-view/survey-view.component';
 import {SurveysComponent} from './views/surveys/surveys.component';
