+++ public-ui/src/app/data-browser/chart/chart.component.html
@@ -1,4 +1,6 @@
-
+<chart class="standard-chart"  [options]="chartOptions"></chart>
+<!-- save instance version for fancy highcharts stuff later
 <chart class="standard-chart"  [options]="chartOptions" (load)="saveInstance($event.context )"></chart>
+-->
