+++ public-ui/src/app/views/pm/pm.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-physical-measurements',
-  templateUrl: './physical-measurements.component.html',
-  styleUrls: ['./physical-measurements.component.css']
+  templateUrl: './pm.component.html',
+  styleUrls: ['./pm.component.css']
 })
 export class PhysicalMeasurementsComponent implements OnInit {
